+++ src/lastversion/__about__.py
@@ -1,4 +1,4 @@
 """Package metadata"""
 
-__version__ = "3.6.2"
+__version__ = "3.6.3"
 __self__ = "dvershinin/lastversion"